chaned how the daemon works so that it only opens the file if there are changes made to the file, also added a 5 second wait after each check as to not tax the cpu

+++ ReminderDaemon.py
@@ -5,36 +5,55 @@ from signal import SIGTERM
 from daemon import Daemon
 from time_converter import TimeConverter
 
-## To Do chang the daemon to only open the file and read in lines if the
-## last modified time is different. 
-
 # Subclass for daemon this will read from the reminder_log
 # read all lines in the reminder log and if the current time
 # equals one of the reminder times send that reminder.
 # then removes that line
 class RDaemon(Daemon):
         def run(self):
-                while True:
-                    try:
-                        f = open(self.logfile, "r+")
-                        lines = f.readlines()
-                        f.seek(0) # put buffer at front
-                        for line in lines:
-                            line_ar = line.split()
-                            if int(line_ar[0]) == int(time.time()):
-                                Notify.init('Remind Me')
-                                notification = Notify.Notification.new('REMINDER', ' '.join(line_ar[1:]))
-                                notification.show()
-                            else:
-                                f.write(line)
-                        f.truncate()
-                        f.close()
-                    except IOError:
-                        print("no reminder log")
-                    # Notify.init('Remind Me')
-                    # notification = Notify.Notification.new('REMINDER', test)
-                    # notification.show()
-                    # time.sleep(10)
+            try:
+                lastModified = int(os.path.getmtime(self.logfile))
+                pastReminders = []
+                f = open(self.logfile, "r+")
+                lines = f.readlines()
+                f.close()
+                Notify.init('Remind Me')
+            except OSError:
+                print("no reminder log")
+                self.stop()
+
+            while True:
+                # if last modified is different than the last time daemon
+                # grabbed the lines we need to update local reminders
+                if lastModified != int(os.path.getmtime(self.logfile)):
+                    f = open(self.logfile, "r+")
+                    lines = f.readlines()
+                    f.close()
+                    lastModified = int(os.path.getmtime(self.logfile))
+
+                # scan all the local reminders to see if one needs to be executed
+                for line in lines:
+                    line_ar = line.split()
+                    if int(line_ar[0]) == int(time.time()):
+                        notification = Notify.Notification.new('REMINDER', ' '.join(line_ar[1:]))
+                        notification.show()
+                        pastReminders.append(line)
+
+                # if we have past reminders clean them up and update local lines
+                if pastReminders:
+                    f = open(self.logfile, "r+")
+                    # lines = f.readlines()
+                    f.seek(0) # put buffer at front
+                    for line in lines:
+                        if line not in pastReminders:
+                            f.write(line)
+                    f.truncate()
+                    lines = f.readlines()
+                    f.close()
+                    lastModified = int(os.path.getmtime(self.logfile))
+                    pastReminders = []
+                time.sleep(5)
+
 
         def stop(self):
                 """
@@ -47,13 +66,13 @@ class RDaemon(Daemon):
                         pf.close()
                 except IOError:
                         pid = None
-        
+
                 if not pid:
                         message = "pidfile %s does not exist. Daemon not running?\n"
                         sys.stderr.write(message % self.pidfile)
                         return # not an error in a restart
-    
-                # Try killing the daemon process       
+
+                # Try killing the daemon process
                 try:
                         while 1:
                                 Notify.uninit()
@@ -66,4 +85,4 @@ class RDaemon(Daemon):
                                         os.remove(self.pidfile)
                         else:
                                 print str(err)
-                                sys.exit(1)        
+                                sys.exit(1)
